+++ R/itpPrint.R
@@ -23,9 +23,10 @@
 #' @export
 print.itp <- function(x, all = FALSE,
                       digits = max(3L, getOption("digits") - 3L), ...) {
-  temp<- x
+  temp <- x
   names(temp) <- c("root", "f(root)", "iterations", "a", "b", "f.a", "f.b",
                    "precision")
+  cat("function:", attr(x, "f_name"), "\n")
   if (!all) {
     print.default(format(temp[1:3], digits = digits), print.gap = 2L,
                   quote = FALSE)