+++ tests/testthat/test-errors.R
@@ -45,3 +45,12 @@ test_that("Lambert: n0 < 0", {
   testthat::expect_error(itp(lambert, c(-1, 1), n0 = -1))
 })
 
+# f not an R function or an external pointer
+test_that("Wrong f", {
+  testthat::expect_error(itp(f = "lambert", c(-1, 1)))
+})
+
+# Trying to create a pointer based on a function that does not exist
+test_that("No C++ function", {
+  testthat::expect_error(xptr_create("fn_does_not_exist"))
+})